refactor: move enum in socket class to CommonEnum.ts
